requirements.txt
astroid==2.0.4
certifi==2018.8.24
colorama==0.3.9
dj-database-url==0.5.0
Django==2.1.1
django-crispy-forms==1.7.2
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
psutil==5.4.7
psycopg2==2.7.5
pytz==2018.5
six==1.11.0
snowballstemmer==1.2.1
typed-ast==1.1.0
whitenoise==4.1
wincertstore==0.2
wrapt==1.10.11
boto3==1.9.16
django-storages==1.7.1
